Add new dependencies for Puppeteer.
This relates to the changes introduced in meteor/meteor#9814, particularly
the new Aptitude packages necessary to support the Puppeteer functionality,
as seen in https://github.com/toinevk/meteor/commit/ee47ff56d87a9ea.

By utilizing this CircleCI image that we've already needed to implement for
Android builds, we'll be able to speed up the initial CircleCI start times
by avoiding the need to install the packages from apt sources on every
build.